add: 상세정보 탭 추가
