+++ frontend/src/components/ProductDetail/ProductMain/ProductTab.tsx
@@ -1,6 +1,6 @@
 import ProductDescription from '@/components/ProductDetail/ProductMain/Detail/ProductDescription.tsx';
 import ProductReview from '@/components/ProductDetail/ProductMain/Review/ProductReview.tsx';
-import ProductDetailInfo from '@/components/ProductDetail/ProductMain/Detail/ProductDetailInfo.tsx';
+import ProductInfo from '@/components/ProductDetail/ProductMain/Detail/ProductInfo.tsx';
 import Loading from '@/components/Common/Loading/Loading.tsx';
 import { Suspense } from 'react';
 import useFetchProductDetail from '@/hooks/fetch/useFetchProductDetail.ts';
@@ -13,7 +13,7 @@ export default function ProductTab({ selectedTab, productId }) {
       <Suspense fallback={<Loading />}>
         {selectedTab === '상품설명' && <ProductDescription data={data} />}
         {selectedTab === '선물후기' && <ProductReview />}
-        {selectedTab === '상세정보' && <ProductDetailInfo announcements={announcements}/>}
+        {selectedTab === '상세정보' && <ProductInfo announcements={announcements}/>}
       </Suspense>
     </>
   );
